Organize extension module a little more
